Incrementos no projeto.

+++ src/main/java/com/example/demo/services/UsuarioService.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -30,12 +31,39 @@ public class UsuarioService {
     }
 
     public Usuario buscarPorId(Long id) {
-        return usuarioRepository.getReferenceById(id);
+        Optional<Usuario> usuarioOptional = usuarioRepository.findById(id);
+        if(usuarioOptional.isEmpty()) {
+            throw new RuntimeException("Usuário não encontrado.");
+        } else {
+            return usuarioOptional.get();
+        }
     }
 
 //    public Usuario atualizar(Long id, Usuario dadosAtualizacao) {}
 
     public void remover(Long id) {
-        usuarioRepository.deleteById(id);
+        Optional<Usuario> usuarioOptional = usuarioRepository.findById(id);
+        if(usuarioOptional.isEmpty()) {
+            throw new RuntimeException("Usuário não encontrado.");
+        } else {
+            usuarioRepository.deleteById(id);
+        }
+    }
+
+    public Usuario atualizarUsuario(Long id, UsuarioDTO usuarioDTO) {
+        if(usuarioDTO.getEmail() != null && !usuarioDTO.getEmail().equals(buscarPorId(id).getEmail())) {
+            buscarPorId(id).setEmail(usuarioDTO.getEmail());
+        }
+        if(usuarioDTO.getNome() != null && !usuarioDTO.getNome().equals(buscarPorId(id).getNome())) {
+            buscarPorId(id).setNome(usuarioDTO.getNome());
+        }
+        if(usuarioDTO.getPais() != null && !usuarioDTO.getPais().equals(buscarPorId(id).getPais())) {
+            buscarPorId(id).setPais(usuarioDTO.getPais());
+        }
+        if(usuarioDTO.getSenha() != null && !usuarioDTO.getSenha().equals(buscarPorId(id).getSenha())) {
+            buscarPorId(id).setSenha(usuarioDTO.getSenha());
+        }
+        usuarioRepository.save(buscarPorId(id));
+        return buscarPorId(id);
     }
 }
